encapsulate video preview comp, removed react-player
