+++ client/src/components/Video/Video.css
@@ -1,0 +1,20 @@
+.Video__Wrapper {
+    float: none;
+    clear: both;
+    width: 100%;
+    position: relative;
+    padding-bottom: 56.25%;
+    padding-top: 25px;
+    height: 0;
+    margin-left:auto; 
+    margin-right:auto;
+
+}
+.Video__Wrapper iframe {
+    position: fixed;
+    top: 25vh;
+    left: 25vw;
+    width: 60vw;
+    height: 60vh;
+    max-width:640px;
+}
